refactor: split test_bundled_recipes.py into 4 focused files + consolidate triplicated tests
Split the 2,387-line monolithic test file into:
- test_bundled_recipes_general.py: top-level invariants, TestBaseBranchDefaults,
  TestImplementationRecipeMergeQueueRule, TestRunModeIngredient
- test_bundled_recipes_pipeline_structure.py: three variant classes + new
  TestPipelineVariantInvariants (16 triplicated methods × 3 params)
- test_bundled_recipes_review_pr.py: TestReviewPrRecipeIntegration (extracted as-is)
- test_bundled_recipes_research.py: TestResearchRecipeStructure (extracted as-is)

Adds tests/arch/test_bundled_recipes_split.py with enforcement guards T1-T4.
Removes test_ip_compose_pr_routes_to_extract_pr_number pre-existing duplicate.
Net test count change: -1 (only the pre-existing duplicate, as planned).
